% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{add_component}
\alias{add_component}
\title{Create a reaction for a pathway}
\usage{
add_component(ip, name, ..., variable = TRUE)
}
\arguments{
\item{name}{of the component (use a vector to add multiple components with the same isotopes and \code{variable} attribute)}

\item{...}{the isotopes (refer to by name)}

\item{variable}{- whether this component has variable concentration and isotopic composition, default is that it can vary}
}
\description{
Create a reaction for a pathway
}
\note{
non-standard evaluation
}

