% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{add_component_}
\alias{add_component_}
\title{add isotopes to a component}
\usage{
add_component_(ip, name, ..., .dots = list(), variable = FALSE)
}
\description{
add isotopes to a component
}
\note{
standard evaluation
}

