% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotope.R
\name{add_isotope}
\alias{add_isotope}
\title{Add an isotope}
\usage{
add_isotope(ip, name)
}
\arguments{
\item{name}{the name of the isotope to add (alpha numeric names only), can be a vector to add multiple isotopes at once}
}
\description{
Add an isotope
}
