% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{add_reaction_}
\alias{add_reaction_}
\title{add reaction with standard evaluation}
\usage{
add_reaction_(ip, name, eq, nr, flux, isotopes = list())
}
\arguments{
\item{nr}{the reaction number (for sorting it in a diagram)}

\item{flux}{lazy object for later evaluation}

\item{isotopes}{named list with lazy objects for later evaluation}
}
\description{
add reaction with standard evaluation
}
\note{
standard evaluation
}

