% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{add_simple_reaction}
\alias{add_simple_reaction}
\title{Add a simple single fractionation reaction}
\usage{
add_simple_reaction(ip, name = NULL, eq, flux = NULL, ...,
  abscissa = NULL)
}
\arguments{
\item{ip}{An \link{isopath} object to which this reaction should be added.}

\item{name}{Name of the reaction. Must be unique (i.e. will overwrite an existing reaction of the same name). If omitted (the default), will name the reaction based on the existing number of reactions in the \link{isopath}, e.g. "rxn4".}

\item{eq}{Valid equation in chemical format a * A + b * B + ... == x * X + y * Y + ... Note that all components must already be part of the \link{isopath} (add components using \link{add_component}).}

\item{flux}{the net flux through the reaction - can be a number or expression (referencing variables and parameters in the system)}

\item{...}{overwrite}

\item{abscissa}{The reaction loaction. This is used purely for graphical representation (see \link{generate_reaction_diagram}). By default, a decent value will be guessed from the existing reactions in the \link{isopath}. This works fairly well as long as all reactions after the first one are added to the isopath in an order such that at least one reactant or product is already part of a different reaction in the isopath.}
}
\description{
This adds a reaction
}
\details{
- check later that fluxes all add up correctly
- check later that the starting point is non-variable - is that actually required? prob not!
}
\seealso{
Other reactions: \code{\link{add_custom_reaction}},
  \code{\link{add_reaction_}}
}

