% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{default_abscissa}
\alias{default_abscissa}
\title{calculate default abscissa
depending on exising reactions, calculate
where the new reaction should go}
\usage{
default_abscissa(ip, new_components)
}
\description{
calculate default abscissa
depending on exising reactions, calculate
where the new reaction should go
}
