% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{dir_flux}
\alias{dir_flux}
\title{Calculate forward/reverse flux}
\usage{
dir_flux(net, reversibility, direction, model_offset = 1e-09)
}
\arguments{
\item{net}{the net flux}

\item{reversibility}{the reversibility of the reaction}

\item{direction}{whether to go forward ("+") or reverse ("-")}

\item{model_offset}{completely reversible reactions (\code{reversibility = 1})
do not have any net flux so if net is != 0 this is a problem that can be solved
at steady-state but not during a model. The \code{model_offset} parameter offsets
completely reversible reactions from 1 by this small number to make the reaction
computationally feasible.}
}
\description{
Convenience function to calculate forward and reverse flux from net flux and
reversibility.
}
\note{
This function uses standard evaluation.
}
