% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{expand_data_frame}
\alias{expand_data_frame}
\title{expand data frame
utility function to expand data frames with non-standard evaluation
used by \code{\link{expand_parameters}}}
\usage{
expand_data_frame(df, ..., .dots = list())
}
\arguments{
\item{...}{use for non standard evaluation}

\item{.dots}{use for standard evaluation}
}
\description{
expand data frame
utility function to expand data frames with non-standard evaluation
used by \code{\link{expand_parameters}}
}

