% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generate_reaction_diagram}
\alias{generate_reaction_diagram}
\title{plot reaction diagram}
\usage{
generate_reaction_diagram(ip, y_data = NULL, add_arrows = FALSE)
}
\arguments{
\item{y_data}{data frame with component and y columns to be used for plotting. If missing, arranges diagram like reaction scheme.}

\item{arrows}{whether to generate arrows for reaction diagram}
}
\description{
plot reaction diagram
}

