% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_component_change_summary}
\alias{get_component_change_summary}
\title{get the summary of component changes (dx/dt) for the parameters provided}
\usage{
get_component_change_summary(ip, parameters = ip$parameters)
}
\description{
get the summary of component changes (dx/dt) for the parameters provided
}

