% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{get_component_matrix}
\alias{get_component_matrix}
\title{get the component matrix for an isopath}
\usage{
get_component_matrix(ip, na = "unspecified")
}
\arguments{
\item{na}{how to group compounds that have no isotopes specified}
}
\description{
get the component matrix for an isopath
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_flux_isotope_matrix}},
  \code{\link{get_flux_matrix}},
  \code{\link{get_isotope_change_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_matrix}}
}

