% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_flux_isotope_matrix}
\alias{get_flux_isotope_matrix}
\title{retrieve flux isotopes matrix}
\usage{
get_flux_isotope_matrix(ip, evaluate = FALSE, parameters = ip$parameters[1,
  ])
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes}

\item{parameters}{only needed if \code{evaluate=TRUE}, data frame with all the parameters needed to evaluate the flux isotope expressions}
}
\description{
retrieve flux isotopes matrix
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_component_matrix}},
  \code{\link{get_flux_matrix}},
  \code{\link{get_isotope_change_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_ode_matrix}},
  \code{\link{get_reaction_component_matrix2}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_isotope_matrix}},
  \code{\link{get_reaction_matrix2}},
  \code{\link{get_reaction_matrix}}
}

