% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_flux_matrix}
\alias{get_flux_matrix}
\title{retrieve flux matrix}
\usage{
get_flux_matrix(ip, evaluate = FALSE, parameters = ip$parameters[1, ])
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes (requires parameters provided)}

\item{parameters}{data to use for evaluating expressions}
}
\description{
retrieve flux matrix
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_component_matrix}},
  \code{\link{get_flux_isotope_matrix}},
  \code{\link{get_isotope_change_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_matrix}}
}

