% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_isotope_change_matrix}
\alias{get_isotope_change_matrix}
\title{get isotope change matrix by reaction and component + isotope}
\usage{
get_isotope_change_matrix(ip, parameters = ip$parameters[1, ],
  check_missing = TRUE)
}
\arguments{
\item{check_missing}{whether to check for any missing flux components}
}
\description{
get isotope change matrix by reaction and component + isotope
}

