% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_isotope_change_summary}
\alias{get_isotope_change_summary}
\title{get the isotope change summary}
\usage{
get_isotope_change_summary(ip, parameters = ip$parameters)
}
\description{
get the isotope change summary
}

