% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{get_reaction_component_matrix}
\alias{get_reaction_component_matrix}
\title{get the combined reaction_component_matrix}
\usage{
get_reaction_component_matrix(ip)
}
\description{
get the combined reaction_component_matrix
}

