% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_reaction_component_matrix2}
\alias{get_reaction_component_matrix2}
\title{get the combined reaction_component_matrix}
\usage{
get_reaction_component_matrix2(ip, evaluate = FALSE,
  parameters = ip$parameters[1, ])
}
\description{
get the combined reaction_component_matrix
}
\note{
TODO: write tests and replace get_reaction_component_matrix
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_component_matrix}},
  \code{\link{get_flux_isotope_matrix}},
  \code{\link{get_flux_matrix}},
  \code{\link{get_isotope_change_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_ode_matrix}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_isotope_matrix}},
  \code{\link{get_reaction_matrix2}},
  \code{\link{get_reaction_matrix}}
}

