% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_reaction_isotope_matrix}
\alias{get_reaction_isotope_matrix}
\title{get the combined reaction_isotope_matrix}
\usage{
get_reaction_isotope_matrix(
  ip,
  evaluate = FALSE,
  parameters = ip$parameters[1, ]
)
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes (requires parameters provided)}

\item{parameters}{data to use for evaluating expressions}
}
\description{
if \code{evaluate = TRUE}, excludes components that are not variable
}
\note{
TODO: write tests and replace all the get_flux... functions!!
}
\seealso{
Other system information: 
\code{\link{get_component_matrix}()},
\code{\link{get_ode_matrix}()},
\code{\link{get_reaction_component_matrix}()},
\code{\link{get_reaction_matrix}()}
}
\concept{system information}
