% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_reaction_matrix}
\alias{get_reaction_matrix}
\title{get the reaction matrix for an isopath}
\usage{
get_reaction_matrix(ip, evaluate = FALSE, parameters = ip$parameters[1, ])
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes (requires parameters provided)}

\item{parameters}{data to use for evaluating expressions}
}
\description{
get the reaction matrix for an isopath
}
\note{
TODO: write tests and replace get_reaction_matrix
}
\seealso{
Other system information: \code{\link{get_component_matrix}},
  \code{\link{get_ode_matrix}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_isotope_matrix}}
}

