% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isocyclr}
\alias{isocyclr}
\title{Model isotope fractionation in biogeochemical cycles and pathways}
\description{
This package facilitates the modeling of isotopic effects through biogeochemical pathways and cycles including non-steady state and steady-state solutions.
}
\author{
Sebastian Kopf
}
