% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{mutate_parameters}
\alias{mutate_parameters}
\title{mutate parameters}
\usage{
mutate_parameters(ip, ...)
}
\arguments{
\item{...}{mutate expressions}
}
\description{
modify existing parameters with mutate statements
}
\seealso{
Other parameters: \code{\link{expand_parameters}},
  \code{\link{set_parameters}}
}

