% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{parse_reaction_component}
\alias{parse_reaction_component}
\title{parse reaction component
pulls out the stoichiometry and component}
\usage{
parse_reaction_component(x)
}
\arguments{
\item{x}{the stoichiometric component, e.g. "A", "2 * A", etc.}
}
\description{
parse reaction component
pulls out the stoichiometry and component
}
\note{
standard evaluation
}

