% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{parse_reaction_equation}
\alias{parse_reaction_equation}
\title{parse reaction equation into its components}
\usage{
parse_reaction_equation(eq)
}
\description{
parse reaction equation into its components
}
\note{
standard evaluation
}
