% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_model}
\alias{run_model}
\title{Run the reaction model}
\usage{
run_model(ip, time_steps, ..., quiet = FALSE)
}
\arguments{
\item{time_steps}{the number of time steps, can be a number or expression (referring to a parameter in the isopath)}

\item{...}{additional parameters passed on to the \link{ode} solver}
}
\description{
Run it for a certain number of time steps.
}
\details{
This will purposefully include the entire parameter set in the state variables to allow for events to change
any parameter at specific time points or in response to functional values
}

