% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_steady_state}
\alias{run_steady_state}
\title{Run model to steady state}
\usage{
run_steady_state(ip, ...)
}
\arguments{
\item{...}{additional parameters passed on to the steady state ODE (\link{runsteady}) solver, use \link{runsteady} parameter \code{verbose = TRUE} for additional output during computations, use parameters \code{rtol}, \code{atol} and \code{ctol} to adjust steady-state tolerances}
}
\description{
Reports the number of time steps required to reach steady-state (time = steps * dt) and creates <variable>.t0 columns for each state variable. Uses \link{runsteady} internally for greatest flexibility although \link{stode} can work for some reaction systems and would be a little faster.
}

