% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{add_component}
\alias{add_component}
\title{Create a reaction for a pathway}
\usage{
add_component(ip, name, ..., constant = FALSE)
}
\arguments{
\item{...}{the isotopes (refer to by name)}

\item{constant}{- whether this component has constant concentration and isotopic composition, default is that it can vary}
}
\description{
Create a reaction for a pathway
}
\note{
non-standard evaluation
}

