% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{add_reaction}
\alias{add_reaction}
\alias{add_reaction_}
\title{add reaction with standard evaluation}
\usage{
add_reaction_(ip, name, eq, flux, isotopes = list(), abscissa = NULL,
  class = "generic", args = list())
}
\arguments{
\item{flux}{lazy object for later evaluation}

\item{isotopes}{named list with lazy objects for later evaluation to calculate the isotope composition of the flux through this reaction}

\item{abscissa}{the reaction loaction (purely for sorting in a digram), a value will be inferred from the system if NULL}

\item{class}{the class of the reaction for internal processing purposes (e.g. steady state calculations)}

\item{args}{additional arguments (usually expressions)}
}
\description{
add reaction with standard evaluation
}
\note{
standard evaluation
}

