% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_component_change_matrix}
\alias{get_component_change_matrix}
\title{get component changes (dx/dt) for each reaction and component with the parameters provided}
\usage{
get_component_change_matrix(ip, parameters = ip$parameters)
}
\description{
get component changes (dx/dt) for each reaction and component with the parameters provided
}

