% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_component_matrix}
\alias{get_component_matrix}
\title{get the component matrix for an isopath}
\usage{
get_component_matrix(ip, na = "unspecified")
}
\arguments{
\item{na}{how to group compounds that have no isotopes specified, if NULL will omit these columns}
}
\description{
get the component matrix for an isopath
}
\seealso{
Other system information: 
\code{\link{get_ode_matrix}()},
\code{\link{get_reaction_component_matrix}()},
\code{\link{get_reaction_isotope_matrix}()},
\code{\link{get_reaction_matrix}()}
}
\concept{system information}
