% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{get_component_matrix}
\alias{get_component_matrix}
\title{get the component matrix for an isopath}
\usage{
get_component_matrix(ip, na = "unspecified")
}
\arguments{
\item{na}{how to group compounds that have no isotopes specified}
}
\description{
get the component matrix for an isopath
}

