% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_flux_isotope_matrix}
\alias{get_flux_isotope_matrix}
\title{retrieve flux isotopes matrix}
\usage{
get_flux_isotope_matrix(ip, evaluate = FALSE, parameters = ip$parameters[1,
  ])
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes}

\item{parameters}{only needed if \code{evaluate=TRUE}, data frame with all the parameters needed to evaluate the flux isotope expressions}
}
\description{
retrieve flux isotopes matrix
}

