% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{get_isotope_change_matrix}
\alias{get_isotope_change_matrix}
\title{get isotope change matrix by reaction and component + isotope}
\usage{
get_isotope_change_matrix(ip, parameters = ip$parameters[1, ],
  check_missing = TRUE)
}
\arguments{
\item{check_missing}{whether to check for any missing flux components}
}
\description{
get isotope change matrix by reaction and component + isotope
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_component_matrix}},
  \code{\link{get_flux_isotope_matrix}},
  \code{\link{get_flux_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_ode_matrix}},
  \code{\link{get_reaction_component_matrix2}},
  \code{\link{get_reaction_component_matrix}},
  \code{\link{get_reaction_isotope_matrix}},
  \code{\link{get_reaction_matrix2}},
  \code{\link{get_reaction_matrix}}
}

