% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{get_ode_matrix}
\alias{get_ode_matrix}
\title{get the ODE}
\usage{
get_ode_matrix(ip, evaluate = FALSE, parameters = ip$parameters[1, ])
}
\arguments{
\item{evaluate}{if TRUE, tries to evaluate the expressions stored for the different fluxes (requires parameters provided)}

\item{parameters}{data to use for evaluating expressions}
}
\description{
Get the system of ordinary differential equations
}
\note{
TODO: write tests and replace all the get_flux... functions!!
}
\seealso{
Other system information: 
\code{\link{get_component_matrix}()},
\code{\link{get_reaction_component_matrix}()},
\code{\link{get_reaction_isotope_matrix}()},
\code{\link{get_reaction_matrix}()}
}
\concept{system information}
