% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction.R
\name{get_reaction_matrix}
\alias{get_reaction_matrix}
\title{get the reaction matrix for an isopath}
\usage{
get_reaction_matrix(ip)
}
\description{
get the reaction matrix for an isopath
}
\seealso{
Other system information: \code{\link{get_component_change_matrix}},
  \code{\link{get_component_change_summary}},
  \code{\link{get_component_matrix}},
  \code{\link{get_flux_isotope_matrix}},
  \code{\link{get_flux_matrix}},
  \code{\link{get_isotope_change_matrix}},
  \code{\link{get_isotope_change_summary}},
  \code{\link{get_reaction_component_matrix}}
}

