% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{set_parameters}
\alias{get_parameters}
\alias{set_parameters}
\title{add parameter data frame to the isopath}
\usage{
set_parameters(ip, ...)

get_parameters(ip)
}
\arguments{
\item{...}{parameters, either a whole set of parameters as a single data frame, or a list of individual values, will try to merge with existing parameters and if neceesary overwrite existing parameters with the same name}
}
\description{
to expand on an existing parameter set with multiple variations, use \code{\link{expand_parameters}}
}
\seealso{
Other parameters: \code{\link{expand_parameters}},
  \code{\link{mutate_parameters}}
}

